+++ src/rules.ts
@@ -17,6 +17,7 @@ import './html/move-style-into-template';
 import './html/undefined-elements';
 
 import './polymer/behaviors-spelling';
+import './polymer/call-super-in-callbacks';
 import './polymer/databind-with-unknown-property';
 import './polymer/element-before-dom-module';
 import './polymer/set-unknown-attribute';